python/learning/06-pytorch/requirements.txt
# Phase 6: PyTorch Dependencies

# Core PyTorch
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Training & Monitoring
tensorboard>=2.15.0
pytorch-lightning>=2.1.0

# Computer Vision
timm>=0.9.0  # PyTorch Image Models
albumentations>=1.3.0  # Data augmentation

# NLP (Optional)
transformers>=4.35.0
tokenizers>=0.15.0

# Model Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
tqdm>=4.66.0
pillow>=10.1.0

# Data Science (from Phase 5)
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.25.0
